move index.html to the static folder
