+++ app/static/index.html
@@ -8,7 +8,7 @@
 
   <meta http-equiv="Content-Security-Policy" content="default-src 'none'; img-src 'self' https: data:; media-src 'none'; child-src 'self'; object-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'; connect-src 'self' https:; font-src 'self';">
 
-  <link rel="stylesheet" href="../static/style/app.css">
+  <link rel="stylesheet" href="./style/app.css">
 </head>
 
 <body>
@@ -42,7 +42,7 @@
 
   </div>
 
-  <script type="text/javascript" src="./index.js"></script>
+  <script type="text/javascript" src="../src/index.js"></script>
 </body>
 
 </html>
