+++ package.json
@@ -2,7 +2,7 @@
   "name": "apparatus",
   "version": "0.1.6",
   "description": "an IoT network security analysis and visualization tool",
-  "main": "./app/src/main.js",
+  "main": "./app/main.js",
   "scripts": {
     "start": "electron .",
     "dev": "cross-env NODE_ENV=development electron .",
